requirement.txt
defusedxml==0.4.1
Django==1.9.7
django-allauth==0.25.2
django-crispy-forms==1.6.0
oauthlib==1.1.2
python3-openid==3.0.10
requests==2.10.0
requests-oauthlib==0.6.1
stripe==1.35.0
